- modified CoverUp evaluation runs to facilitate re-starting, picking up from   where it left, based on its checkpoint, and retaining all files no matter   how the process ends;
- added --interactive to facilitate debugging/poking around interactively;

- added pip-cache to retain a local pip cache across docker executions,
  speeding them up;

+++ eval/run_coverup.sh
@@ -9,48 +9,63 @@ run() {
     eval $1 || exit 1
 }
 
-run "cp -R /package ."
+cd /output
+
+# create writable version of the package
+if ! [ -d package ]; then
+    run "cp -R /package ."
+
+    # requirementslib has a variable named _fragment_dict and pydantic, if also installed,
+    # causes load failures (saying it should be renamed to fragment_dict).
+    run "sed -i '/requirementslib/d' package/package.txt"
+    run "chown -R $OWNER /output"
+fi
+
 run "cd package"
 
-# requirementslib has a variable named _fragment_dict and pydantic, if also installed,
-# causes load failures (saying it should be renamed to fragment_dict).
-run "sed -i '/requirementslib/d' package.txt"
-
 run "pip install -r package.txt || true"    # ignore any errors because so did CodaMOSA
-run "pip install -r /coverup/requirements.txt"
-run "pip install -r /coverup/test-modules.txt"
+run "pip install -r /coverup/requirements.txt -r /coverup/test-modules.txt"
 
 PYTEST_ARGS="--rootdir . -c /dev/null" # ignore configuration which would deviate from expected defaults
 SLIPCOVER_ARGS="--source $SRC/$PKG --branch --json"
+COVERUP_ARGS="--write-requirements-to /output/requirements.txt --log-file /output/coverup-log --source-dir $SRC/$PKG --tests-dir coverup-tests --pytest-args \"$PYTEST_ARGS\""
+
+[ -d coverup-tests ] || mkdir coverup-tests
 
 # generate initial coverage file where nothing is covered
-mkdir coverup-tests
-cat >coverup-tests/test_dummy.py <<XXX
+if ! [ -e /output/initial.json ]; then
+    cat >coverup-tests/test_dummy.py <<XXX
 def test_dummy():
     pass
 XXX
-run "python3 -m slipcover $SLIPCOVER_ARGS --out /output/initial.json -m pytest $PYTEST_ARGS coverup-tests"
-run "chown -R $OWNER /output"
+
+    run "python3 -m slipcover $SLIPCOVER_ARGS --out /output/initial.json -m pytest $PYTEST_ARGS coverup-tests"
+    rm coverup-tests/test_dummy.py
+
+    run "chown -R $OWNER /output"
+fi
 
 # set Python path so the package can be 'import'ed
 export PYTHONPATH=$SRC
 
-# run CoverUp on it
-COVERUP_ARGS="--no-checkpoint --write-requirements-to /output/requirements.txt --log-file /output/coverup-log --source-dir $SRC/$PKG --tests-dir coverup-tests --pytest-args \"$PYTEST_ARGS\""
-run "python3 /coverup/coverup.py $COVERUP_ARGS /output/initial.json $FILES"
-run "chown -R $OWNER /output"
+# 1st pass
+if ! [ -e /output/interim.json ]; then
+    # run CoverUp on it
+    run "python3 /coverup/coverup.py $COVERUP_ARGS --no-initial-test-check /output/initial.json $FILES"
+    run "mv coverup-ckpt.json coverup-ckpt-0.json"
 
-# measure where coverage's at
-run "rm coverup-tests/test_dummy.py"
-run "python3 -m slipcover $SLIPCOVER_ARGS --out /output/interim.json -m pytest $PYTEST_ARGS coverup-tests"
-run "chown -R $OWNER /output"
+    # measure where coverage's at
+    run "python3 -m slipcover $SLIPCOVER_ARGS --out /output/interim.json -m pytest $PYTEST_ARGS coverup-tests"
 
-# re-run CoverUp to try to improve on it
-echo "-------- 2nd run --------" >> /output/coverup-log
-run "python3 /coverup/coverup.py $COVERUP_ARGS /output/interim.json $FILES"
-run "cp -R coverup-tests /output"
-run "chown -R $OWNER /output"
+    run "chown -R $OWNER /output"
+fi
 
-# measure final coverage
-run "python3 -m slipcover $SLIPCOVER_ARGS --out /output/final.json -m pytest $PYTEST_ARGS coverup-tests"
-run "chown -R $OWNER /output"
+# 2nd pass: re-run to try to improve coverage
+if ! [ -e /output/final.json ]; then
+    run "python3 /coverup/coverup.py $COVERUP_ARGS /output/interim.json $FILES"
+
+    # measure final coverage
+    run "python3 -m slipcover $SLIPCOVER_ARGS --out /output/final.json -m pytest $PYTEST_ARGS coverup-tests"
+
+    run "chown -R $OWNER /output"
+fi
